update: 適切な粒度でコンポーネントを分割
