+++ src/app/interfaces/IProps.ts
@@ -5,8 +5,21 @@ export interface CheckBoxPrefecturesProps {
   prefectures: PrefecturesData[];
 }
 
+export interface CheckBoxInputProps {
+  onChange: (prefCode: number, prefName: string, isChecked: boolean) => void;
+  prefCode: number;
+  prefName: string;
+}
+
 export interface CheckBoxPrefPopulationProps {
   labels: string[];
+  onChange: (index: number) => void;
+}
+
+export interface RadioInputProps {
+  defaultChecked?: boolean;
+  index: number;
+  label: string;
   onChange: (index: number) => void;
 }
